--- a/packages/g6/src/types/graph.ts
+++ b/packages/g6/src/types/graph.ts
@@ -470,7 +470,11 @@ export interface IGraph<
    * @returns
    * @group Item
    */
-  hideItem: (ids: ID | ID[], disableAnimate?: boolean) => void;
+  hideItem: (
+    ids: ID | ID[],
+    disableAnimate?: boolean,
+    keepKeyShape?: boolean,
+  ) => void;
   /**
    * Make the item(s) to the front.
    * @param ids the item id(s) to front