--- a/packages/g6/src/types/hook.ts
+++ b/packages/g6/src/types/hook.ts
@@ -53,6 +53,7 @@ export interface Hooks {
     graphCore: GraphCore;
     theme: ThemeSpecification;
     transientCanvas: Canvas;
+    tileFirstRender?: number | boolean;
   }>; // TODO: define param template
   layout: IHook<{
     graphCore: GraphCore;
@@ -82,6 +83,7 @@ export interface Hooks {
     action?: string;
     enableStack?: boolean;
     changes?: any;
+    keepKeyShape?: boolean;
   }>;
   itemzindexchange: IHook<{
     ids: ID[];