… packages/g6/src/stdlib/behavior/drag-canvas.ts
… packages/g6/src/stdlib/behavior/zoom-canvas.ts
… packages/g6/src/stdlib/item/edge/base.ts
… packages/g6/src/stdlib/item/node/base.ts
… packages/g6/src/types/spec.ts
… packages/g6/src/util/bbox.ts
… packages/site/examples/scatter/changePosition/demo/meta.json