feat: tile-based rendering for first rendering and drag-canvas zoom-canvas optimization;
Graph 可配置 optimize: { tileFirstRender: number | false, behavior: number | false }  来控制首次渲染和交互的分片渲染优化。
支持分片渲染，提升首屏渲染以及拖拽画布、缩放画布中涉及到全量图形的绘制、更新。

perf: performance of label size keeping;
通过分片更新，提升缩放过程中字号保持的效率

效果：
- 首屏幕渲染 5.5w 点边 webgl，白屏时间 7s -> 4s
- 拖拽画布、缩放画布 1.5w 点边 canvas，20FPS -> 30FPS